requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
pycodestyle==2.7.0
requests==2.25.1
toml==0.10.2
torch==1.8.1+cpu
torchvision==0.9.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
